Move plugin runtime into plugin-system package
