+++ ui/plugin-system/src/index.ts
@@ -13,3 +13,4 @@
 
 export * from './components';
 export * from './model';
+export * from './runtime';
