+++ ui/plugin-system/src/runtime/template-variables.ts
